Add reasons why I created this to ReadMe.md
this repository is not meant to be offensive but to be a reminder to be kind to others. you can be mad, but that doesn't change the truth. sure, this language is extremely intense, offensive, and vulgar but it gets a point across rapidly. 
do not be mad at the messenger but the person who sent the site to you.