add files to make the website existent in the interwebs
this repository is not meant to be offensive but to be a reminder to be kind to others. you can be mad, but that doesn't change the truth. sure, this language is extremely intense, offensive, and vulgar but it gets a point across rapidly. do not be mad at the messenger but the person who sent the site to you.

send me hate mail 😍: i would love to read it – i appreciate the time you will be taking to write it

if you want to support me 🙌🏻 – why not help me pay for college? don't want to do that – tell me to send it to a charity – i will. any amount accepted just follow the link – cash.me/$donutblikethat

hopefully this will help give you another reason to be kind to people around you – cheers! – donutblikethat